+++ app/about/about.module.css
@@ -1,0 +1,8 @@
+.bio {
+  min-height: 133px;
+}
+
+.button {
+  font-size: 1.2rem;
+  padding: 10px;
+}
