+++ app/about/page.module.css
@@ -1,0 +1,4 @@
+.container {
+  padding: 10px 25px;
+  max-width: 675px;
+}
